moved dem validation functions to use json instead of umm_json
