+++ api/src/hyp3_api/util.py
@@ -11,10 +11,6 @@ class QuotaError(Exception):
     pass
 
 
-class CmrError(Exception):
-    pass
-
-
 def format_time(time: datetime):
     if time.tzinfo is None:
         raise ValueError(f'missing tzinfo for datetime {time}')
@@ -47,20 +43,3 @@ def get_request_time_expression(start, end):
         return key.gte(formatted_start)
     if formatted_end:
         return key.lte(formatted_end)
-
-
-def check_granules_exist(granules):
-    cmr_parameters = {
-        'producer_granule_id': granules,
-        'provider': 'ASF',
-        'short_name': [
-            'SENTINEL-1A_SLC',
-            'SENTINEL-1B_SLC',
-        ],
-    }
-    response = requests.post(CMR_URL, data=cmr_parameters)
-    response.raise_for_status()
-    found_granules = [entry['producer_granule_id'] for entry in response.json()['feed']['entry']]
-    not_found_granules = set(granules) - set(found_granules)
-    if not_found_granules:
-        raise CmrError(f'Some requested scenes could not be found: {",".join(not_found_granules)}')
